Win32: Adjust type casting to meet VC++ reqs
VC++ doesn't allow pointer arithmetic on void*. Instead, we cast to
char* so it does the arithmetic in terms of byte units.

This commit also fixes a few other places where char* needs to be cast
to void*, or vice-versa, as well as the integer types use for some
loops.

+++ DocFormats/lib/DFBuffer.c
@@ -118,7 +118,7 @@ DFBuffer *DFBufferReadFromFile(const char *filename, DFError **error)
         return NULL;
     }
     DFBuffer *buf = DFBufferNew();
-    ssize_t r;
+    long r;
     char temp[1024];
     while (0 < (r = read(fd,temp,1024)))
         DFBufferAppendData(buf,temp,r);
